Replace libstandard with object libraries and add a stdcompiler target
There is no need to build a static library where object libraries suffice.
Also, a separate stdcompiler target, building a shared library, is added,
in case anyone wants to use it in a separate project.

Note that standard and stdcompiler aren't built with C4ENGINE defined anymore -
it wasn't used before anyway (except for an unused include in Standard.cpp),
and whoever needs to differentiate between clonk and c4group shouldn't add it
to standard anyway.